refactor: extract background processing into stage module and slim down inbound handler
