+++ assistant/src/runtime/routes/inbound-message-handler.ts
@@ -6,14 +6,12 @@
  */
 // Side-effect imports: register channel invite transport adapters so the
 // ACL enforcement module can resolve transports at runtime.
-import type { ChannelId, InterfaceId } from "../../channels/types.js";
 import {
   CHANNEL_IDS,
   INTERFACE_IDS,
   isChannelId,
   parseInterfaceId,
 } from "../../channels/types.js";
-import { resolveUserReference } from "../../config/user-reference.js";
 import type { TrustContext } from "../../daemon/session-runtime-assembly.js";
 import * as attachmentsStore from "../../memory/attachments-store.js";
 import {
@@ -32,11 +30,6 @@ import { IngressBlockedError } from "../../util/errors.js";
 import { getLogger } from "../../util/logger.js";
 import { DAEMON_INTERNAL_ASSISTANT_ID } from "../assistant-scope.js";
 import { mintDaemonDeliveryToken } from "../auth/token-service.js";
-import {
-  buildApprovalUIMetadata,
-  getApprovalInfoByConversation,
-  getChannelApprovalPrompt,
-} from "../channel-approvals.js";
 import { getGuardianBinding } from "../channel-guardian-service.js";
 import { deliverChannelReply } from "../gateway-client.js";
 import { routeGuardianReply } from "../guardian-reply-router.js";
@@ -48,18 +41,14 @@ import type {
   GuardianFollowUpConversationGenerator,
   MessageProcessor,
 } from "../http-types.js";
-import {
-  resolveRoutingState,
-  resolveTrustContext,
-} from "../trust-context-resolver.js";
-import { deliverReplyViaCallback } from "./channel-delivery-routes.js";
+import { resolveTrustContext } from "../trust-context-resolver.js";
 import {
   canonicalChannelAssistantId,
   GUARDIAN_APPROVAL_TTL_MS,
 } from "./channel-route-shared.js";
 import { handleApprovalInterception } from "./guardian-approval-interception.js";
-import { deliverGeneratedApprovalPrompt } from "./guardian-approval-prompt.js";
 import { enforceIngressAcl } from "./inbound-stages/acl-enforcement.js";
+import { processChannelMessageInBackground } from "./inbound-stages/background-dispatch.js";
 import { handleBootstrapIntercept } from "./inbound-stages/bootstrap-intercept.js";
 import { handleVerificationIntercept } from "./inbound-stages/verification-intercept.js";
 
@@ -875,440 +864,4 @@ export async function handleChannelInbound(
     duplicate: result.duplicate,
     eventId: result.eventId,
   });
-}
-
-// ---------------------------------------------------------------------------
-// Background message processing
-// ---------------------------------------------------------------------------
-
-interface BackgroundProcessingParams {
-  processMessage: MessageProcessor;
-  conversationId: string;
-  eventId: string;
-  content: string;
-  attachmentIds?: string[];
-  sourceChannel: ChannelId;
-  sourceInterface: InterfaceId;
-  externalChatId: string;
-  trustCtx: TrustContext;
-  metadataHints: string[];
-  metadataUxBrief?: string;
-  replyCallbackUrl?: string;
-  /** Factory that mints a fresh delivery JWT for each HTTP attempt. */
-  mintBearerToken: () => string;
-  assistantId?: string;
-  approvalCopyGenerator?: ApprovalCopyGenerator;
-  commandIntent?: Record<string, unknown>;
-  sourceLanguageCode?: string;
-}
-
-const TELEGRAM_TYPING_INTERVAL_MS = 4_000;
-const PENDING_APPROVAL_POLL_INTERVAL_MS = 300;
-
-// Module-level map tracking which approval requestIds have already been
-// notified to trusted contacts. Maps requestId -> conversationId so that
-// cleanup can be scoped to the owning conversation's poller, preventing
-// concurrent pollers from different conversations from evicting each
-// other's entries.
-const globalNotifiedApprovalRequestIds = new Map<string, string>();
-
-function delay(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-function shouldEmitTelegramTyping(
-  sourceChannel: ChannelId,
-  replyCallbackUrl?: string,
-): boolean {
-  if (sourceChannel !== "telegram" || !replyCallbackUrl) return false;
-  try {
-    return new URL(replyCallbackUrl).pathname.endsWith("/deliver/telegram");
-  } catch {
-    return replyCallbackUrl.endsWith("/deliver/telegram");
-  }
-}
-
-function startTelegramTypingHeartbeat(
-  callbackUrl: string,
-  chatId: string,
-  mintBearerToken: () => string,
-  assistantId?: string,
-): () => void {
-  let active = true;
-  let inFlight = false;
-
-  const emitTyping = (): void => {
-    if (!active || inFlight) return;
-    inFlight = true;
-    void deliverChannelReply(
-      callbackUrl,
-      { chatId, chatAction: "typing", assistantId },
-      mintBearerToken(),
-    )
-      .catch((err) => {
-        log.debug(
-          { err, chatId },
-          "Failed to deliver Telegram typing indicator",
-        );
-      })
-      .finally(() => {
-        inFlight = false;
-      });
-  };
-
-  emitTyping();
-
-  const interval = setInterval(emitTyping, TELEGRAM_TYPING_INTERVAL_MS);
-  (interval as { unref?: () => void }).unref?.();
-
-  return () => {
-    active = false;
-    clearInterval(interval);
-  };
-}
-
-function startPendingApprovalPromptWatcher(params: {
-  conversationId: string;
-  sourceChannel: ChannelId;
-  externalChatId: string;
-  trustClass: TrustContext["trustClass"];
-  guardianExternalUserId?: string;
-  requesterExternalUserId?: string;
-  replyCallbackUrl: string;
-  mintBearerToken: () => string;
-  assistantId?: string;
-  approvalCopyGenerator?: ApprovalCopyGenerator;
-}): () => void {
-  const {
-    conversationId,
-    sourceChannel,
-    externalChatId,
-    trustClass,
-    guardianExternalUserId,
-    requesterExternalUserId,
-    replyCallbackUrl,
-    mintBearerToken,
-    assistantId,
-    approvalCopyGenerator,
-  } = params;
-
-  // Approval prompt delivery is guardian-only. Non-guardian and unverified
-  // actors must never receive approval prompt broadcasts for the conversation.
-  // We also require an explicit identity match against the bound guardian to
-  // avoid broadcasting prompts when trustClass is stale/mis-scoped.
-  const isBoundGuardianActor =
-    trustClass === "guardian" &&
-    !!guardianExternalUserId &&
-    requesterExternalUserId === guardianExternalUserId;
-  if (!isBoundGuardianActor) {
-    return () => {};
-  }
-
-  let active = true;
-  const deliveredRequestIds = new Set<string>();
-
-  const poll = async (): Promise<void> => {
-    while (active) {
-      try {
-        const prompt = getChannelApprovalPrompt(conversationId);
-        const pending = getApprovalInfoByConversation(conversationId);
-        const info = pending[0];
-        if (prompt && info && !deliveredRequestIds.has(info.requestId)) {
-          deliveredRequestIds.add(info.requestId);
-          const delivered = await deliverGeneratedApprovalPrompt({
-            replyCallbackUrl,
-            chatId: externalChatId,
-            sourceChannel,
-            assistantId: assistantId ?? DAEMON_INTERNAL_ASSISTANT_ID,
-            bearerToken: mintBearerToken(),
-            prompt,
-            uiMetadata: buildApprovalUIMetadata(prompt, info),
-            messageContext: {
-              scenario: "standard_prompt",
-              toolName: info.toolName,
-              channel: sourceChannel,
-            },
-            approvalCopyGenerator,
-          });
-          if (!delivered) {
-            // Delivery can fail transiently (network or gateway outage).
-            // Keep polling and retry prompt delivery for the same request.
-            deliveredRequestIds.delete(info.requestId);
-          }
-        }
-      } catch (err) {
-        log.warn(
-          { err, conversationId },
-          "Pending approval prompt watcher failed",
-        );
-      }
-      await delay(PENDING_APPROVAL_POLL_INTERVAL_MS);
-    }
-  };
-
-  void poll();
-  return () => {
-    active = false;
-  };
-}
-
-/**
- * Resolve a human-readable guardian name from the guardian binding metadata.
- * Returns the display name, username (prefixed with @), or undefined if
- * no name is available.
- */
-function resolveGuardianDisplayName(
-  assistantId: string,
-  sourceChannel: ChannelId,
-): string | undefined {
-  const binding = getGuardianBinding(assistantId, sourceChannel);
-  if (!binding?.metadataJson) return undefined;
-  try {
-    const parsed = JSON.parse(binding.metadataJson) as Record<string, unknown>;
-    if (
-      typeof parsed.displayName === "string" &&
-      parsed.displayName.trim().length > 0
-    ) {
-      return parsed.displayName.trim();
-    }
-    if (
-      typeof parsed.username === "string" &&
-      parsed.username.trim().length > 0
-    ) {
-      return `@${parsed.username.trim()}`;
-    }
-  } catch {
-    // ignore malformed metadata
-  }
-  return undefined;
-}
-
-/**
- * Start a poller that sends a one-shot "waiting for guardian approval" message
- * to the trusted contact when a confirmation_request enters guardian approval
- * wait. Deduplicates by requestId so each request only produces one message.
- *
- * Only activates for trusted-contact actors with a resolvable guardian route.
- */
-function startTrustedContactApprovalNotifier(params: {
-  conversationId: string;
-  sourceChannel: ChannelId;
-  externalChatId: string;
-  trustClass: TrustContext["trustClass"];
-  guardianExternalUserId?: string;
-  replyCallbackUrl: string;
-  mintBearerToken: () => string;
-  assistantId?: string;
-}): () => void {
-  const {
-    conversationId,
-    sourceChannel,
-    externalChatId,
-    trustClass,
-    guardianExternalUserId,
-    replyCallbackUrl,
-    mintBearerToken,
-    assistantId,
-  } = params;
-
-  // Only notify trusted contacts who have a resolvable guardian route.
-  if (trustClass !== "trusted_contact" || !guardianExternalUserId) {
-    return () => {};
-  }
-
-  let active = true;
-
-  const poll = async (): Promise<void> => {
-    while (active) {
-      try {
-        const pending = getApprovalInfoByConversation(conversationId);
-        const info = pending[0];
-
-        // Clean up resolved requests from the module-level dedupe map.
-        // Only remove entries that belong to THIS conversation — other
-        // conversations' pollers own their own entries. Without this
-        // scoping, concurrent pollers would evict each other's request
-        // IDs and cause duplicate notifications.
-        const currentPendingIds = new Set(pending.map((p) => p.requestId));
-        for (const [rid, cid] of globalNotifiedApprovalRequestIds) {
-          if (cid === conversationId && !currentPendingIds.has(rid)) {
-            globalNotifiedApprovalRequestIds.delete(rid);
-          }
-        }
-
-        if (info && !globalNotifiedApprovalRequestIds.has(info.requestId)) {
-          globalNotifiedApprovalRequestIds.set(info.requestId, conversationId);
-          const guardianName =
-            resolveGuardianDisplayName(
-              assistantId ?? DAEMON_INTERNAL_ASSISTANT_ID,
-              sourceChannel,
-            ) ?? resolveUserReference();
-          const waitingText = `Waiting for ${guardianName}'s approval...`;
-          try {
-            await deliverChannelReply(
-              replyCallbackUrl,
-              {
-                chatId: externalChatId,
-                text: waitingText,
-                assistantId: assistantId ?? DAEMON_INTERNAL_ASSISTANT_ID,
-              },
-              mintBearerToken(),
-            );
-          } catch (err) {
-            log.warn(
-              { err, conversationId },
-              "Failed to deliver trusted-contact pending-approval notification",
-            );
-            // Remove from notified set so delivery is retried on next poll
-            globalNotifiedApprovalRequestIds.delete(info.requestId);
-          }
-        }
-      } catch (err) {
-        log.warn(
-          { err, conversationId },
-          "Trusted-contact approval notifier poll failed",
-        );
-      }
-      await delay(PENDING_APPROVAL_POLL_INTERVAL_MS);
-    }
-  };
-
-  void poll();
-  return () => {
-    active = false;
-
-    // Evict all dedupe entries owned by this conversation so the
-    // module-level map doesn't grow unboundedly after the poller stops.
-    for (const [rid, cid] of globalNotifiedApprovalRequestIds) {
-      if (cid === conversationId) {
-        globalNotifiedApprovalRequestIds.delete(rid);
-      }
-    }
-  };
-}
-
-function processChannelMessageInBackground(
-  params: BackgroundProcessingParams,
-): void {
-  const {
-    processMessage,
-    conversationId,
-    eventId,
-    content,
-    attachmentIds,
-    sourceChannel,
-    sourceInterface,
-    externalChatId,
-    trustCtx,
-    metadataHints,
-    metadataUxBrief,
-    replyCallbackUrl,
-    mintBearerToken,
-    assistantId,
-    approvalCopyGenerator,
-    commandIntent,
-    sourceLanguageCode,
-  } = params;
-
-  (async () => {
-    const typingCallbackUrl = shouldEmitTelegramTyping(
-      sourceChannel,
-      replyCallbackUrl,
-    )
-      ? replyCallbackUrl
-      : undefined;
-    const stopTypingHeartbeat = typingCallbackUrl
-      ? startTelegramTypingHeartbeat(
-          typingCallbackUrl,
-          externalChatId,
-          mintBearerToken,
-          assistantId,
-        )
-      : undefined;
-    const stopApprovalWatcher = replyCallbackUrl
-      ? startPendingApprovalPromptWatcher({
-          conversationId,
-          sourceChannel,
-          externalChatId,
-          trustClass: trustCtx.trustClass,
-          guardianExternalUserId: trustCtx.guardianExternalUserId,
-          requesterExternalUserId: trustCtx.requesterExternalUserId,
-          replyCallbackUrl,
-          mintBearerToken,
-          assistantId,
-          approvalCopyGenerator,
-        })
-      : undefined;
-    const stopTcApprovalNotifier = replyCallbackUrl
-      ? startTrustedContactApprovalNotifier({
-          conversationId,
-          sourceChannel,
-          externalChatId,
-          trustClass: trustCtx.trustClass,
-          guardianExternalUserId: trustCtx.guardianExternalUserId,
-          replyCallbackUrl,
-          mintBearerToken,
-          assistantId,
-        })
-      : undefined;
-
-    try {
-      const cmdIntent =
-        commandIntent && typeof commandIntent.type === "string"
-          ? {
-              type: commandIntent.type as string,
-              ...(typeof commandIntent.payload === "string"
-                ? { payload: commandIntent.payload }
-                : {}),
-              ...(sourceLanguageCode
-                ? { languageCode: sourceLanguageCode }
-                : {}),
-            }
-          : undefined;
-      const { messageId: userMessageId } = await processMessage(
-        conversationId,
-        content,
-        attachmentIds,
-        {
-          transport: {
-            channelId: sourceChannel,
-            hints: metadataHints.length > 0 ? metadataHints : undefined,
-            uxBrief: metadataUxBrief,
-          },
-          assistantId,
-          trustContext: trustCtx,
-          isInteractive: resolveRoutingState(trustCtx).promptWaitingAllowed,
-          ...(cmdIntent ? { commandIntent: cmdIntent } : {}),
-        },
-        sourceChannel,
-        sourceInterface,
-      );
-      channelDeliveryStore.linkMessage(eventId, userMessageId);
-      channelDeliveryStore.markProcessed(eventId);
-
-      if (replyCallbackUrl) {
-        await deliverReplyViaCallback(
-          conversationId,
-          externalChatId,
-          replyCallbackUrl,
-          mintBearerToken(),
-          assistantId,
-          {
-            onSegmentDelivered: (count) =>
-              channelDeliveryStore.updateDeliveredSegmentCount(eventId, count),
-          },
-        );
-      }
-    } catch (err) {
-      log.error(
-        { err, conversationId },
-        "Background channel message processing failed",
-      );
-      channelDeliveryStore.recordProcessingFailure(eventId, err);
-    } finally {
-      stopTypingHeartbeat?.();
-      stopApprovalWatcher?.();
-      stopTcApprovalNotifier?.();
-    }
-  })();
 }
